+++ admin/src/new-test/new.ts
@@ -41,3 +41,5 @@ function generateTestConfigs() {
   }];
   globalThis.configs = newConfigs;
 }
+
+