+++ functions_backup/main.py
@@ -14,6 +14,7 @@ from psp_liquids_daq_parser import (
 )
 from google.cloud.storage import transfer_manager
 import requests
+from createTest_helpers import downloadFromGDrive, getUnits, organizeFiles
 from firebase_functions.firestore_fn import (
     on_document_created,
     Event,
@@ -21,46 +22,13 @@ from firebase_functions.firestore_fn import (
 )
 import h5py
 from azure.core.exceptions import ResourceExistsError, ResourceNotFoundError
-from datetime import datetime
-import time as t
-import gdown
+
 from azure.storage.fileshare import ShareFileClient
 
 app: App = initialize_app()
 storage_client = storage
 db = firestore.client()
 
-
-def organizeFiles(file_names: list[str]):
-    csv_files = list(filter(lambda x: ".csv" in x, file_names))
-    fileNames = list(filter(lambda x: ".csv" not in x, file_names))
-    fileNames.sort()
-    timestamps: list[int] = []
-    for file in fileNames:
-        time_stamp_str = file[8:25]
-        datetimeObj = datetime.strptime(time_stamp_str, "%Y-%m%d-%H%M-%S")
-        dateString = t.mktime(datetimeObj.timetuple())
-        timestamps.append(int(dateString))
-    return (fileNames, csv_files, timestamps)
-
-
-def downloadFromGDrive(url: str):
-    print("new download: " + url)
-    file_name = gdown.download(url=url, fuzzy=True)
-    return file_name
-
-
-def getUnits(dataset_name: str) -> str:
-    scale = "psi"
-    if "tc" in dataset_name:
-        scale = "deg"
-    if "pi-" in dataset_name or "reed-" in dataset_name or "_state" in dataset_name:
-        scale = "bin"
-    if "fms" in dataset_name:
-        scale = "lbf"
-    if "rtd" in dataset_name:
-        scale = "V"
-    return scale
 
 @https_fn.on_call(timeout_sec=540, memory=options.MemoryOption.GB_8, cpu=2)
 def createTest_links_gdrive(req: https_fn.CallableRequest) -> Any: